@@ -8,8 +8,8 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1",
-        "spatie/laravel-permission": "^2.16"
+        "php": "^7.4|^8.0",
+        "spatie/laravel-permission": "^3.0|^4.0|^5.0"
     },
     "autoload": {
         "psr-4": {
@@ -18,7 +18,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0-dev"
+            "dev-master": "2.0-dev"
         },
         "laravel": {
             "providers": [
