@@ -2,10 +2,9 @@
 
 namespace Vyuldashev\NovaPermission;
 
-use Gate;
-use Laravel\Nova\Nova;
-use Laravel\Nova\Events\ServingNova;
 use Illuminate\Support\ServiceProvider;
+use Laravel\Nova\Events\ServingNova;
+use Laravel\Nova\Nova;
 
 class ToolServiceProvider extends ServiceProvider
 {
@@ -16,23 +15,15 @@ class ToolServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        $this->loadViewsFrom(__DIR__.'/../resources/views', 'nova-permission-tool');
         $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'nova-permission-tool');
 
         $this->publishes([
             __DIR__.'/../resources/lang' => resource_path('lang/vendor/nova-permission-tool'),
         ], 'nova-permission-tool-lang');
 
-        $this->publishes([
-            __DIR__.'/../resources/views' => resource_path('views/vendor/nova-permission-tool'),
-        ], 'nova-permission-tool-views');
-
         $this->app->booted(function () {
             $this->routes();
         });
-
-        Gate::policy(config('permission.models.permission'), PermissionPolicy::class);
-        Gate::policy(config('permission.models.role'), RolePolicy::class);
 
         Nova::serving(function (ServingNova $event) {
             //
